+++ CHANGELOG.md
@@ -3,11 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.4.5
+## 3.4.4.6
 
 ### New Features
 
 ### Fixes
-- Reinstalling Vanilla instances with different Minecraft versions without a loader causing error [#480]
 
 ### Misc